requirements.txt
coreapi
coreschema
Django
django-cors-headers
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow
psycopg2-binary
PyJWT
python-decouple
requests
django-cors-headers
gunicorn
django-notifications-hq